chore(plugins): pull code that can be re-used across multiple component major versions into the main `component` module
