+++ crates/zeroclaw-plugins/src/component/mod.rs
@@ -4,7 +4,10 @@
 // `ComponentChannel`) and the shared `ComponentEngine`. The `bindings` module
 // is internal — consumers use the adapters instead of the bindgen output.
 
+mod call_plugin;
 mod engine;
+pub(crate) mod plugin_store;
 pub mod v0;
+mod wrap_plugin;
 
 pub use engine::ComponentEngine;
