+++ crates/zeroclaw-plugins/src/component/v0/memory_component.rs
@@ -26,8 +26,9 @@ use super::bindings::memory::{
         ProceduralMessage as WitProceduralMessage,
     },
 };
-use super::plugin_store::{self, PluginStore};
+use super::plugin_linker;
 use crate::component::engine::ComponentEngine;
+use crate::component::plugin_store::PluginStore;
 use crate::error::PluginError;
 use crate::{FineGrainedPermission, call_plugin};
 
@@ -76,7 +77,7 @@ impl ComponentMemory {
         wasmtime_wasi::p2::add_to_linker_async(&mut linker).map_err(PluginError::from)?;
         wasmtime_wasi_http::p2::add_only_http_to_linker_async(&mut linker)
             .map_err(PluginError::from)?;
-        plugin_store::add_to_linker_memory(&mut linker)?;
+        plugin_linker::add_to_linker_memory(&mut linker)?;
         let host = PluginStore::with_permissions(permissions, &network_config).await?;
         let mut store = wasmtime::Store::new(engine.engine(), host);
