+++ crates/zeroclaw-plugins/src/component/wrap_plugin.rs
